+++ ui/src/app/shared/layout/sidebar/sidebar.component.ts
@@ -12,7 +12,6 @@ import { AuthService } from '@/app/core/auth/auth.service';
 import { InformationComponent } from '@/app/core/components/information/information.component';
 import { NotificationService } from '@/app/core/notification.service';
 import { SettingsService } from '@/app/core/settings.service';
-import { PowerOptionsComponent } from '@/app/shared/layout/power-options/power-options.component';
 
 @Component({
   selector: 'app-sidebar',
@@ -60,16 +59,6 @@ export class SidebarComponent implements OnInit {
     this.toggleSidebar.emit(!this.isExpanded);
   }
 
-  openRestartModal() {
-    if (this.isExpanded) {
-      this.toggleSidebar.emit(false);
-    }
-    this.$modal.open(PowerOptionsComponent, {
-      size: 'lg',
-      backdrop: 'static',
-    });
-  }
-
   openUnderVoltageModal() {
     const ref = this.$modal.open(InformationComponent, {
       size: 'lg',
@@ -84,5 +73,14 @@ export class SidebarComponent implements OnInit {
     ref.componentInstance.ctaButtonLabel = this.$translate.instant('form.button_more_info');
     ref.componentInstance.faIconClass = 'fas fa-fw fa-bolt yellow-text';
     ref.componentInstance.ctaButtonLink = 'https://pimylifeup.com/raspberry-pi-low-voltage-warning';
+  }
+
+  handleKeydown(event: KeyboardEvent) {
+    if (event.key === 'Enter') {
+      const target = event.target as HTMLElement;
+      if (['menuitem', 'button'].includes(target.getAttribute('role'))) {
+        target.click();
+      }
+    }
   }
 }